refactoring the project as a django app template
