+++ app_name/package.json
@@ -1,0 +1,14 @@
+{
+  "name": "Wheelie",
+  "version": "0.0.0",
+  "description": "Django Wheelie",
+  "main": "gulpfile.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "devDependencies": {
+    "gulp": "3.9.0",
+    "wheelie": "0.1.3",
+    "wheelie-recipe": "0.2.1"
+  }
+}
