--- frontend/client/js/main.js
@@ -1,13 +1,0 @@
-window.MyModule = (function() {
-    'use strict';
-
-    function MyModule() {
-        // noop
-    }
-
-    MyModule.prototype.foo = function() {
-        console.log("bar");
-    };
-
-    return MyModule;
-})();
