--- frontend/client/scss/main.scss
@@ -1,5 +1,0 @@
-$background: #333333;
-
-body {
-    background-color: $background;
-}
