--- frontend/gulpfile.js
@@ -1,28 +1,0 @@
-require('gulp');
-
-// importing Wheelie instance and a list of tasks (recipe)
-var wheelie = require('wheelie');
-var recipe = require('wheelie-recipe');
-
-// adding a recipe to Wheelie, defining the default task
-wheelie.add(recipe);
-wheelie.setDefault('watch');
-
-// build customizations
-var vendors = [
-    // put here your vendors installed via bower (or whatever), available in `client/vendors/` folder
-];
-
-var scripts = [
-    'main.js'
-    // add here your JS files available in `client/js/` folder
-];
-
-wheelie.update('uglify', {
-  scripts: scripts,
-  vendors: vendors
-});
-
-wheelie.setBuild('static/');
-wheelie.setDist('static/');
-wheelie.build();
