--- frontend/package.json
@@ -1,16 +1,0 @@
-{
-  "name": "frontend",
-  "version": "0.0.0",
-  "description": "Django frontend",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "Emanuele Palazzetti",
-  "license": "Free-BSD-2",
-  "devDependencies": {
-    "gulp": "^3.9.0",
-    "wheelie": "^0.1.3",
-    "wheelie-recipe": "^0.2.1"
-  }
-}
