--- frontend/static/css/main.css
@@ -1,5 +1,0 @@
-/* line 3, stdin */
-body {
-  background-color: #333333; }
-
-/*# sourceMappingURL=data:application/json;base64,eyJ2ZXJzaW9uIjozLCJzb3VyY2VzIjpbIm1haW4uY3NzIiwibWFpbi5zY3NzIl0sIm5hbWVzIjpbXSwibWFwcGluZ3MiOiJBQUFBLG1CQUFtQjtBQ0VuQjtFQUNJLDBCQUhnQixFQUVkIiwiZmlsZSI6Im1haW4uY3NzIiwic291cmNlc0NvbnRlbnQiOlsiLyogbGluZSAzLCBzdGRpbiAqL1xuYm9keSB7XG4gIGJhY2tncm91bmQtY29sb3I6ICMzMzMzMzM7IH1cbiIsIiRiYWNrZ3JvdW5kOiAjMzMzMzMzO1xuXG5ib2R5IHtcbiAgICBiYWNrZ3JvdW5kLWNvbG9yOiAkYmFja2dyb3VuZDtcbn1cbiJdLCJzb3VyY2VSb290IjoiL3NvdXJjZS8ifQ== */
