--- frontend/static/js/app.js
@@ -1,15 +1,0 @@
-window.MyModule = (function() {
-    'use strict';
-
-    function MyModule() {
-        // noop
-    }
-
-    MyModule.prototype.foo = function() {
-        console.log("bar");
-    };
-
-    return MyModule;
-})();
-
-//# sourceMappingURL=data:application/json;base64,eyJ2ZXJzaW9uIjozLCJzb3VyY2VzIjpbImNsaWVudC9qcy9tYWluLmpzIl0sIm5hbWVzIjpbXSwibWFwcGluZ3MiOiJBQUFBO0FBQ0E7QUFDQTtBQUNBO0FBQ0E7QUFDQTtBQUNBO0FBQ0E7QUFDQTtBQUNBO0FBQ0E7QUFDQTtBQUNBO0FBQ0EiLCJmaWxlIjoiYXBwLmpzIiwic291cmNlc0NvbnRlbnQiOlsid2luZG93Lk15TW9kdWxlID0gKGZ1bmN0aW9uKCkge1xuICAgICd1c2Ugc3RyaWN0JztcblxuICAgIGZ1bmN0aW9uIE15TW9kdWxlKCkge1xuICAgICAgICAvLyBub29wXG4gICAgfVxuXG4gICAgTXlNb2R1bGUucHJvdG90eXBlLmZvbyA9IGZ1bmN0aW9uKCkge1xuICAgICAgICBjb25zb2xlLmxvZyhcImJhclwiKTtcbiAgICB9O1xuXG4gICAgcmV0dXJuIE15TW9kdWxlO1xufSkoKTtcbiJdLCJzb3VyY2VSb290IjoiL3NvdXJjZS8ifQ==
