--- frontend/views.py
@@ -1,5 +1,0 @@
-from django.views.generic import TemplateView
-
-
-class HomeView(TemplateView):
-    template_name = 'frontend/frontend.html'
