--- requirements.txt
@@ -1,7 +1,0 @@
-#
-# This file is autogenerated by pip-compile
-# Make changes in requirements.in, then run this to update:
-#
-#    pip-compile requirements.in
-#
-django==1.8.4
